minio_api-main/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "minio_api"
dynamic = ["version"]
description = "MinIO-based stock data distribution API"
authors = [{name="Leo", email="luoxin@zoyutech.com"}]
dependencies = [
    "pandas>=1.5.0",
    "minio>=7.0.0",
    "pyarrow>=10.0.0",
    "fastapi>=0.95.0",
    "uvicorn>=0.20.0",
    "python-dotenv",
    "duckdb"
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools_scm]
